feat: user session is now terminated when loggin out
