--- src/main/java/com/espe/zonarbol/routes/LoginSRV.java
@@ -1,35 +1,0 @@
-package com.espe.zonarbol.routes;
-
-import com.espe.zonarbol.service.LoginService;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-
-import java.io.IOException;
-
-@WebServlet("/login")
-public class LoginSRV extends HttpServlet {
-
-    private final LoginService loginService = new LoginService();
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-
-        String username = request.getParameter("username");
-        String password = request.getParameter("password");
-
-        boolean isAuthenticated = loginService.authenticate(username, password);
-
-        if (isAuthenticated) {
-            HttpSession session = request.getSession();
-            session.setAttribute("username", username);
-            response.sendRedirect("menu.jsp");
-        } else {
-            response.sendRedirect("index.jsp?error=1");
-        }
-    }
-}
